+++ tests/hil/scripts/pytest-hil/pytest_hil/nrf52840dk.py
@@ -1,5 +1,5 @@
-from zephyrboard import ZephyrBoard
-from nordicboard import NordicBoard
+from pytest_hil.zephyrboard import ZephyrBoard
+from pytest_hil.nordicboard import NordicBoard
 
 class nRF52840DK(ZephyrBoard, NordicBoard):
     @property
