+++ tests/hil/scripts/pytest-hil/pytest_hil/nrf9160dk.py
@@ -1,5 +1,5 @@
-from ncsboard import NCSBoard
-from nordicboard import NordicBoard
+from pytest_hil.ncsboard import NCSBoard
+from pytest_hil.nordicboard import NordicBoard
 
 class nRF9160DK(NCSBoard, NordicBoard):
     @property
