+++ tests/hil/scripts/pytest-zephyr-samples/pytest_zephyr_samples/plugin.py
